# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AcadVault API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /chat/conversations:
    get:
      operationId: listConversations
      tags: [chat]
      summary: Get user's conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationListResponse"
    post:
      operationId: createConversation
      tags: [chat]
      summary: Start a conversation with another user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateConversationBody"
      responses:
        "201":
          description: Conversation created or existing returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
  /chat/conversations/{conversationId}/messages:
    get:
      operationId: listMessages
      tags: [chat]
      summary: Get messages in a conversation
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: integer
        - name: cursor
          in: query
          required: false
          schema:
            type: integer
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageListResponse"
    post:
      operationId: sendMessage
      tags: [chat]
      summary: Send a message in a conversation
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageBody"
      responses:
        "201":
          description: Sent message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
...
components:
  schemas:
...
    PostAuthor:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        profileImageUrl:
          type: string
      required:
        - id
...
    Conversation:
      type: object
      properties:
        id:
          type: integer
        participant:
          $ref: "#/components/schemas/PostAuthor"
        lastMessage:
          $ref: "#/components/schemas/Message"
        unreadCount:
          type: integer
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - participant
        - unreadCount
        - updatedAt
    ConversationListResponse:
      type: object
      properties:
        conversations:
          type: array
          items:
            $ref: "#/components/schemas/Conversation"
      required:
        - conversations
    CreateConversationBody:
      type: object
      properties:
        userId:
          type: string
      required:
        - userId
    Message:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        senderId:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - senderId
        - content
        - createdAt
    MessageListResponse:
      type: object
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/Message"
        nextCursor:
          type: integer
      required:
        - messages
    SendMessageBody:
      type: object
      properties:
        content:
          type: string
      required:
        - content
...