# lib/api-zod/src/generated/types/postAuthor.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AcadVault API
 * OpenAPI spec version: 0.1.0
 */

export interface PostAuthor {
  id: string;
  username?: string;
  firstName?: string;
  lastName?: string;
  profileImageUrl?: string;
}